requirements.txt
anthropic>=0.3.0
openai>=1.0.0
requests>=2.31.0
replicate>=0.25.0
openai-whisper>=20231117
edge-tts>=6.1.9
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pyperclip>=1.8.0
mutagen>=1.47.0
Pillow>=10.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.7
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.23.0
google-cloud-texttospeech>=2.31.0
google-cloud-aiplatform>=1.38.0
schedule>=1.2.0
pytest>=8.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psutil>=5.9.0
pydantic>=2.5.0
